Move bucket properties operator code into base class preparing for bucket type operator.
